Added pagination and changed the way parameters are processed from request
Too many parameters being passed from request, so restructured the way
parameters are processed so it's done in one place not multiple.
Added pagination function, but still need to write tests for this.
Changed how media type passed around - it was narrowing media type to
RDF type on the ApiService classes, but this was making it more
difficult to tailor response by type in some places. Now media type
passed into response managers.